cmd: cancel context on signal
before this commit, we were already doing the right thing creating a
context with an associated cancel function

	ctx, cancel := context.WithCancel(context.Background())
	defer cancel()

but that will not by default lead to a proper smooth termination when a signal
is sent (either us sending a SIGINT / SIGTERM to make run, or kubelet trying to
stop the pod with SIGINT)

	> By default, a synchronous signal is converted into a run-time panic.
	> A SIGHUP, SIGINT, or SIGTERM signal causes the program to exit.

i.e., despite passing the context to the right places, we're not being graceful
at all (we can notice that that's the case by sending a sigint/sigterm
to the controller and noticing that it just "dies" with no graceful
messaging).

making use of controller-runtime's `SetupSignalHandler` though, we can
have a signal handler that will first call the cancel method (thus, let
things terminate gracefully) and only then terminate.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>